webpack config logic changed in order to handle env variable properly.
