+++ webpack/webpack.dev.config.js
@@ -1,0 +1,3 @@
+module.exports = (coreConfig) => {
+  return coreConfig;
+};
